chore: wip, simple model
