+++ features/sketch/src/main/java/com/wire/android/feature/sketch/DrawingCanvasComponent.kt
@@ -21,8 +21,6 @@ import androidx.compose.foundation.Canvas
 import androidx.compose.foundation.background
 import androidx.compose.foundation.gestures.awaitEachGesture
 import androidx.compose.foundation.gestures.awaitFirstDown
-import androidx.compose.foundation.gestures.calculatePan
-import androidx.compose.foundation.gestures.calculateZoom
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
@@ -32,13 +30,12 @@ import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.Path
 import androidx.compose.ui.graphics.drawscope.Stroke
-import androidx.compose.ui.graphics.drawscope.withTransform
 import androidx.compose.ui.graphics.nativeCanvas
 import androidx.compose.ui.input.pointer.pointerInput
 import androidx.compose.ui.input.pointer.positionChange
-import androidx.compose.ui.layout.onSizeChanged
-import androidx.compose.ui.unit.toSize
 import androidx.lifecycle.viewmodel.compose.viewModel
+import com.wire.android.feature.sketch.model.MotionEvent
+import com.wire.android.feature.sketch.model.PathProperties
 
 @Composable
 fun DrawingCanvasComponent(
@@ -46,92 +43,71 @@ fun DrawingCanvasComponent(
 ) {
     val drawModifier = Modifier
         .fillMaxSize()
-        .clipToBounds()
+        .clipToBounds() // necessary to draw in the canvas.
         .background(MaterialTheme.colorScheme.background)
-        .onSizeChanged {
-            viewModel.canvasState.size = it.toSize()
-            viewModel.canvasState.translation = Offset(it.width / 2f, it.height / 2f)
-        }
         .pointerInput(Unit) {
             awaitEachGesture {
                 val downEvent = awaitFirstDown()
-                viewModel.currentPosition =
-                    (downEvent.position - viewModel.canvasState.translation) / viewModel.canvasState.scale
+                viewModel.currentPosition = downEvent.position - viewModel.canvasState.translation
                 viewModel.motionEvent = MotionEvent.Down
                 if (downEvent.pressed != downEvent.previousPressed) downEvent.consume()
-                var canvasMoved = false
                 do {
                     val event = awaitPointerEvent()
-                    if (event.changes.size == 1) {
-                        if (canvasMoved) break
-                        viewModel.currentPosition =
-                            (event.changes[0].position - viewModel.canvasState.translation) / viewModel.canvasState.scale
-                        viewModel.motionEvent = MotionEvent.Move
-                        if (event.changes[0].positionChange() != Offset.Zero) event.changes[0].consume()
-                    } else if (event.changes.size > 1) {
-                        val zoom = event.calculateZoom()
-                        viewModel.canvasState.scale = (viewModel.canvasState.scale * zoom).coerceIn(0.5f..2f)
-                        val pan = event.calculatePan()
-                        viewModel.canvasState.translation += pan
-                        canvasMoved = true
-                    }
+                    viewModel.currentPosition = event.changes[0].position - viewModel.canvasState.translation
+                    viewModel.motionEvent = MotionEvent.Move
+                    if (event.changes[0].positionChange() != Offset.Zero) event.changes[0].consume()
                 } while (event.changes.any { it.pressed })
                 viewModel.motionEvent = MotionEvent.Up
             }
         }
     Canvas(modifier = drawModifier) {
-        withTransform({
-            translate(viewModel.canvasState.translation.x, viewModel.canvasState.translation.y)
-            scale(viewModel.canvasState.scale, viewModel.canvasState.scale, viewModel.canvasState.pivot)
-        }) {
-            with(drawContext.canvas.nativeCanvas) {
-                val checkPoint = saveLayer(null, null)
-                when (viewModel.motionEvent) {
-                    MotionEvent.Idle -> Unit
-                    MotionEvent.Down -> {
-                        viewModel.paths.add(viewModel.currentPath)
-                        viewModel.currentPath.path.moveTo(
-                            viewModel.currentPosition.x,
-                            viewModel.currentPosition.y
-                        )
-                    }
-
-                    MotionEvent.Move -> {
-                        viewModel.currentPath.path.lineTo(
-                            viewModel.currentPosition.x,
-                            viewModel.currentPosition.y
-                        )
-                        drawCircle(
-                            center = viewModel.currentPosition,
-                            color = Color.Gray,
-                            radius = viewModel.currentPath.strokeWidth / 2,
-                            style = Stroke(
-                                width = 1f
-                            )
-                        )
-                    }
-
-                    MotionEvent.Up -> {
-                        viewModel.currentPath.path.lineTo(
-                            viewModel.currentPosition.x,
-                            viewModel.currentPosition.y
-                        )
-                        viewModel.currentPath = PathProperties(
-                            path = Path(),
-                            strokeWidth = viewModel.currentPath.strokeWidth,
-                            color = viewModel.currentPath.color,
-                            drawMode = viewModel.currentPath.drawMode
-                        )
-                        viewModel.pathsUndone.clear()
-                        viewModel.currentPosition = Offset.Unspecified
-                        viewModel.motionEvent = MotionEvent.Idle
-                    }
+        with(drawContext.canvas.nativeCanvas) {
+            val checkPoint = saveLayer(null, null)
+            when (viewModel.motionEvent) {
+                MotionEvent.Idle -> Unit
+                MotionEvent.Down -> {
+                    viewModel.paths.add(viewModel.currentPath)
+                    viewModel.currentPath.path.moveTo(
+                        viewModel.currentPosition.x,
+                        viewModel.currentPosition.y
+                    )
                 }
-                viewModel.paths.forEach { path ->
-                    path.draw(this@withTransform /*, bitmap*/)
+
+                MotionEvent.Move -> {
+                    viewModel.currentPath.path.lineTo(
+                        viewModel.currentPosition.x,
+                        viewModel.currentPosition.y
+                    )
+                    drawCircle(
+                        center = viewModel.currentPosition,
+                        color = Color.Gray,
+                        radius = viewModel.currentPath.strokeWidth / 2,
+                        style = Stroke(
+                            width = 1f
+                        )
+                    )
                 }
-                restoreToCount(checkPoint)
+
+                MotionEvent.Up -> {
+                    viewModel.currentPath.path.lineTo(
+                        viewModel.currentPosition.x,
+                        viewModel.currentPosition.y
+                    )
+                    viewModel.currentPath = PathProperties(
+                        path = Path(),
+                        strokeWidth = viewModel.currentPath.strokeWidth,
+                        color = viewModel.currentPath.color,
+                        drawMode = viewModel.currentPath.drawMode
+                    )
+                    viewModel.pathsUndone.clear()
+                    viewModel.currentPosition = Offset.Unspecified
+                    viewModel.motionEvent = MotionEvent.Idle
+                }
             }
+            viewModel.paths.forEach { path ->
+                path.draw(this@Canvas /*, bitmap*/)
+            }
+            restoreToCount(checkPoint)
         }
     }
