+++ features/sketch/src/main/java/com/wire/android/feature/sketch/model/MotionEvent.kt
@@ -15,8 +15,8 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see http://www.gnu.org/licenses/.
  */
-package com.wire.android.feature.sketch
+package com.wire.android.feature.sketch.model
 
-enum class DrawMode {
-    Pen, Eraser, None
+enum class MotionEvent {
+    Idle, Down, Move, Up
 }
